Add gallery, Implements #87

+++ src/dgenies/bin/clean_jobs.py
@@ -10,7 +10,7 @@ import argparse
 
 from dgenies.config_reader import AppConfigReader
 from dgenies.lib.functions import Functions
-from dgenies.database import Job
+from dgenies.database import Job, Gallery
 
 config_reader = AppConfigReader()
 
@@ -35,6 +35,7 @@ def parse_upload_folders(upload_folder, now, max_age, fake=False):
 
 
 def parse_database(app_data, max_age, fake=False):
+    gallery_jobs = []
     with Job.connect():
         old_jobs = Job.select().where(
                       ((Job.status == "success") & (Job.date_created < datetime.now() - timedelta(days=max_age["data"])))
@@ -43,6 +44,9 @@ def parse_database(app_data, max_age, fake=False):
                   )
         for job in old_jobs:
             id_job = job.id_job
+            is_gallery = len(Gallery.select().join(Job).where(Job.id_job == id_job)) > 0
+            if is_gallery:
+                gallery_jobs.append(id_job)
             print("Removing job %s..." % id_job)
             data_dir = os.path.join(app_data, id_job)
             if os.path.exists(data_dir) and os.path.isdir(data_dir):
@@ -52,39 +56,41 @@ def parse_database(app_data, max_age, fake=False):
                 print("Job %s has no data folder!" % id_job)
             if not fake:
                 job.delete_instance()
+    return gallery_jobs
 
 
-def parse_data_folders(app_data, now, max_age, fake=False):
+def parse_data_folders(app_data, now, max_age, gallery_jobs, fake=False):
     for file in os.listdir(app_data):
-        file = os.path.join(app_data, file)
-        create_date = os.path.getctime(file)
-        age = (now - create_date) / 86400  # Age in days
-        if age > max_age["data"]:
-            try:
-                if os.path.isdir(file):
-                    print("Removing folder %s..." % file)
-                    if not fake:
-                        shutil.rmtree(file)
-                else:
-                    print("Removing file %s..." % file)
-                    if not fake:
-                        os.remove(file)
-            except OSError:
-                print(traceback.print_exc())
-        elif os.path.isdir(file):
-            query_name_file = os.path.join(file, ".query")
-            if os.path.exists(query_name_file):
-                with open(query_name_file) as query_file:
-                    sorted_file = Functions.get_fasta_file(file, "query", True)
-                    if not sorted_file.endswith(".sorted"):
-                        sorted_file = None
-                    if sorted_file is not None:
-                        create_date = os.path.getctime(sorted_file)
-                        age = (now - create_date) / 86400  # Age in days
-                        if age > max_age["fasta_sorted"]:
-                            print("Removing fasta file %s..." % sorted_file)
-                            if not fake:
-                                os.remove(sorted_file)
+        if file not in gallery_jobs:
+            file = os.path.join(app_data, file)
+            create_date = os.path.getctime(file)
+            age = (now - create_date) / 86400  # Age in days
+            if age > max_age["data"]:
+                try:
+                    if os.path.isdir(file):
+                        print("Removing folder %s..." % file)
+                        if not fake:
+                            shutil.rmtree(file)
+                    else:
+                        print("Removing file %s..." % file)
+                        if not fake:
+                            os.remove(file)
+                except OSError:
+                    print(traceback.print_exc())
+            elif os.path.isdir(file):
+                query_name_file = os.path.join(file, ".query")
+                if os.path.exists(query_name_file):
+                    with open(query_name_file) as query_file:
+                        sorted_file = Functions.get_fasta_file(file, "query", True)
+                        if not sorted_file.endswith(".sorted"):
+                            sorted_file = None
+                        if sorted_file is not None:
+                            create_date = os.path.getctime(sorted_file)
+                            age = (now - create_date) / 86400  # Age in days
+                            if age > max_age["fasta_sorted"]:
+                                print("Removing fasta file %s..." % sorted_file)
+                                if not fake:
+                                    os.remove(sorted_file)
 
 
 if __name__ == '__main__':
@@ -122,7 +128,7 @@ if __name__ == '__main__':
     print("# Parsing Jobs in DB #")
     print("######################")
     print("")
-    parse_database(
+    gallery_jobs = parse_database(
         app_data=app_data,
         max_age=max_age,
         fake=fake
@@ -137,6 +143,7 @@ if __name__ == '__main__':
         app_data=app_data,
         now=now,
         max_age=max_age,
-        fake=fake
+        fake=fake,
+        gallery_jobs=gallery_jobs
     )
     print("")
